refactor: moves sentry out of startup to main
